Make the read syscall work correctly
When the input buffer is empty, the read syscall must return if it has
already read at least one byte, and must wait for more input if it has
not yet read any byte.

Emscripten's `read` syscall calls `stream.tty.ops.get_char` repeatedly
until the specified length of bytes is read or `get_char` returns null.

https://github.com/emscripten-core/emscripten/blob/353fffeef2e6dfaefcd03f6d9c93866b610bd929/src/library_tty.js#L58-L81

Our `emscriptenHack` function overwrites the `get_char` function. So,
when the input buffer is empty, it should wait for input if it is the
first call to `get_char` in the current read syscall, and return null
if it is the second or subsequent call.

However, `get_char` did not know whether it is the first call or not.
This changeset intecepts `TTY.stream_ops.read` to set the flag so that
`get_char` can determine if it is called the first time.

Fixes #3